Maze traversal using DFS and BFS
in this code I have written code to implement a maze structure and i have tested that the maze structure will be perfect for upto 33 x 33. The maze is designed in such a way that each cell contains a cell number which can be used to track the route the DFS and BFS algorithm is taking for traversing the path from the source to the destination.